backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
networkx==3.3
python-dotenv==1.0.1
pydantic==2.9.2
google-generativeai==0.8.3
